requirements.txt
click==7.1.2
Flask==1.1.4
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==2.0.1
marshmallow==3.2.1
mysqlclient==2.1.1
protobuf==3.20.1
pycodestyle==2.9.1
python-decouple==3.1
python-dotenv==0.21.0
SQLAlchemy==1.4.41
toml==0.10.2
Werkzeug==1.0.1